updated for downsampled files

+++ ngsplot.sh
@@ -22,7 +22,23 @@ do
 	esac
 done
 
-if [[ -f ngsplot.conf ]]; 
+if [[ -f downsample/scaling_factors.txt ]]; 
+then
+	sed '1d' downsample/scaling_factors.txt > downsample/scaling_factors-temp.txt #removes header
+	input="downsample/scaling_factors-temp.txt"
+	count=0
+	while IFS= read -r line
+	do
+		input=$(echo $line | cut -d "," -f 2)
+		chip=$(echo $line | cut -d "," -f 4)
+		scaling_output_name=$(echo $line | cut -d "," -f 3)
+		scaling_output_name=${scaling_output_name##*/}
+		scaling_output_name=${scaling_output_name%-ds.bam}
+		ngs.plot.r -G $ngsplot_genome -R tss -C "$input:$chip" -O "ngsplot/$scaling_output_name/${scaling_output_name}_tss" -T $scaling_output_name -L 5000 2>> ngsplot.log
+	done < "$input"
+
+
+if [[ -f ngsplot.conf ]] && [[ ! -f downsample/scaling_factors.txt ]]; 
 then
 	sed '1d' ngsplot.conf > ngsplot-temp.conf #removes header
 	input="ngsplot-temp.conf"
@@ -35,6 +51,7 @@ then
 		ngs.plot.r -G $ngsplot_genome -R tss -C "$ngsplot_chip:$ngsplot_input" -O "ngsplot/$ngsplot_output_name/${ngsplot_output_name}_tss" -T $ngsplot_output_name -L 5000 2>> ngsplot.log
 	done < "$input"
 fi
+
 rm ngsplot-temp.conf *.cnt
 if [[ $# == 1 ]];
 	then
